+++ src/main/java/com/example/bank_app/repository/BankTransactionRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface BankTransactionRepository extends JpaRepository <BankTransaction, Integer> {
+public interface BankTransactionRepository extends JpaRepository <BankTransaction, Long> {
     boolean existsByReferenceCode(String referenceCode);
 
     @Query("SELECT t FROM BankTransaction t WHERE t.sourceAccount.id = :accountId OR t.targetAccount.id = :accountId ORDER BY t.createdAt DESC")
     // Equivale a:
     // @Query(value = "SELECT * FROM bank_transaction WHERE source_account_id = :accountId OR target_account_id = :accountId ORDER BY created_at DESC", nativeQuery = true)
-    List<BankTransaction> findAllByAccountId(@Param("accountId") Integer accountId);
+    List<BankTransaction> findAllByAccountId(@Param("accountId") Long accountId);
 }
