+++ src/main/java/com/example/bank_app/repository/BeneficiaryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BeneficiaryRepository extends JpaRepository<Beneficiary, Integer> {
-    List<Beneficiary> findAllByUserIdAndIsActiveTrue(Integer userId);
-    boolean existsByUserIdAndAccountNumberAndIsActiveTrue(Integer userId, String accountNumber);
+public interface BeneficiaryRepository extends JpaRepository<Beneficiary, Long> {
+    List<Beneficiary> findAllByUserIdAndIsActiveTrue(Long userId);
+    boolean existsByUserIdAndAccountNumberAndIsActiveTrue(Long userId, String accountNumber);
 }
